Fix 32-bit Windows build problems, CentOS build problem, MacOS isCaseSensitive methods.
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@11781 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ interpreter/parser/ProgramSource.cpp
@@ -750,12 +750,12 @@ BufferClass* FileProgramSource::readProgram(const char *file_name)
     size_t readSize;
 
     // get a buffer object to return the image
-    Protected<BufferClass> buffer = new_buffer(bufferSize);
+    Protected<BufferClass> buffer = new_buffer((size_t)bufferSize);
     {
         UnsafeBlock releaser;
 
         // read the data
-        programFile.read(buffer->getData(), bufferSize, readSize);
+        programFile.read(buffer->getData(), (size_t)bufferSize, readSize);
         programFile.close();
     }
     // if there was a read error, return nothing
